+++ src/Arnoldi.hpp
@@ -14,11 +14,11 @@ public:
     inline uint64 rows() const { return n; }
     inline uint64 cols() const { return n; }
     // y = G * G' * x ; data.G is n x m;
-    void perform_op(const float *x_in, float* y_out);
+    void perform_op(const double *x_in, double* y_out);
     inline void setFlags(bool is_update, bool is_standardize, bool is_pcangsd) {update = is_update; standardize = is_standardize; pcangsd = is_pcangsd;}
 
-    MatrixXf U, VT;
-    VectorXf S;
+    MatrixXd U, VT;
+    VectorXd S;
     uint nops;
 
 private:
