+++ src/FilePlink.hpp
@@ -12,8 +12,8 @@
 * 11 ->  0 (copy of A1)
 * 01 ->  3 (missing)
 */
-const float BED_MISSING_VALUE = -9;
-const float BED2GENO[4] = {1, BED_MISSING_VALUE, 0.5, 0};
+const double BED_MISSING_VALUE = -9;
+const double BED2GENO[4] = {1, BED_MISSING_VALUE, 0.5, 0};
 
 class FileBed : public Data
 {
@@ -46,7 +46,7 @@ public:
     virtual void read_all_and_centering();
     // for blockwise
     virtual void read_snp_block_initial(uint64 start_idx, uint64 stop_idx, bool standardize = false);
-    virtual void read_snp_block_update(uint64 start_idx, uint64 stop_idx, const MatrixXf& U, const VectorXf& svals, const MatrixXf& VT, bool standardize = false);
+    virtual void read_snp_block_update(uint64 start_idx, uint64 stop_idx, const MatrixXd& U, const VectorXd& svals, const MatrixXd& VT, bool standardize = false);
 
 
     virtual void check_file_offset_first_var();
